after review: merged formats into manager, improved coverage to 97%
