+++ pass_import/formats/__init__.py
@@ -4,4 +4,3 @@ from pass_import.formats.kdbx import KDBX
 from pass_import.formats.otp import OTP
 from pass_import.formats.xml import HTML, XML
 from pass_import.formats.yaml import YAML
-from pass_import.formats.sphinx import SPHINX
